RavenDB-12161 Calculator from static to instance, terminate process condition changed, CpuUsage to CpuHelper and clean up

+++ src/Raven.Server/Dashboard/ThreadsInfo.cs
@@ -14,7 +14,7 @@ namespace Raven.Server.Dashboard
 
         public double CpuUsage { get; set; }
 
-        public double ActiveCores { get; set; }
+        public long ActiveCores { get; set; }
 
         public ThreadsInfo()
         {
